+++ Missionframework/Extensions/Enemy_Mines/cfgFunctions.hpp
@@ -1,0 +1,6 @@
+class Enemy_Mines
+{
+    file = "Extensions\Enemy_Mines\functions";
+    class createMines {};
+    class createSavedMines {};
+};