+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Added Extension: Commander can edit FOB/Outpost names.
 * Added Extension: Discord log for FOB resources (Added by Fernandim).
 * Added Extension: UI jammer effects for drones near enemy towers.
+* Added Extension: Enemy mines in sectors.
 * Added: Players can build Outposts.
 * Added: `get_presets.sqf` file. It reads the presets selected on mission parameters.
 * Added: Readme file for presets to help mission makers understand them.